+++ CHANGE_LOG.md
@@ -1,8 +1,9 @@
 ### Openapply CHANGE LOG
 
 * **v0.3.1** - compatible with 0.3.x - 2018-02-XX
+  - add error / exception when ENV doesn't load properly
   - remove unused gems (scp and ssh)
-  - add example code
+  - add example / demo usage code
 
 * **v0.3.0** - compatible with 0.3.x - 2018-01-31
   - NOT compatible with 0.2.x